xboxkrnl: Xbox/Windows define DWORD as unsigned long

+++ lib/xboxkrnl/xboxdef.h
@@ -17,7 +17,7 @@ typedef long long LONGLONG, *PLONGLONG;
 typedef unsigned char UCHAR, *PUCHAR;
 typedef unsigned short USHORT, *PUSHORT, CSHORT;
 typedef unsigned short WORD, WCHAR, *PWSTR;
-typedef unsigned int DWORD, *PDWORD, *LPDWORD;
+typedef unsigned long DWORD, *PDWORD, *LPDWORD;
 typedef unsigned long ULONG, *PULONG;
 typedef unsigned long long ULONGLONG;
